Made a simple blockout for level 1
I made a simple block out for level 1. I figured it would be cool to start the story off in a prologue, where the player is trying to broadcast an distress signal from an unknown planet. The player will enter a locked monolithic-like room, and they will have to defeat some form of boss before they're transported to the second level, which is TBD.